Transferrred the classes OccupancyPatch and TSDFPatch from SurfacePatches.hpp tp their own headers. Swapped affected includes in other headers and added the new headers to the CMakeList.txt.
