+++ src/grid/OccupancyPatch.hpp
@@ -24,39 +24,81 @@
 // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-#pragma once
+#ifndef __MAPS_OCCUPANCY_PATCH_HPP_
+#define __MAPS_OCCUPANCY_PATCH_HPP_
 
-#include "OccupancyPatch.hpp"
+
+#include <cmath>
 
 #include <boost/serialization/access.hpp>
 #include <boost/serialization/nvp.hpp>
 
-namespace maps { namespace grid
-{
 
-struct OccupancyConfiguration
-{
-    OccupancyConfiguration(double hit_probability = 0.7, double miss_probability = 0.4,
-                         double occupied_probability = 0.8, double free_space_probability = 0.3,
-                         double max_probability = 0.971, double min_probability = 0.1192,
-                         float uncertainty_threshold = 0.25) :
-                        hit_logodds(OccupancyPatch::logodds(hit_probability)),
-                        miss_logodds(OccupancyPatch::logodds(miss_probability)),
-                        occupied_logodds(OccupancyPatch::logodds(occupied_probability)),
-                        free_space_logodds(OccupancyPatch::logodds(free_space_probability)),
-                        max_logodds(OccupancyPatch::logodds(max_probability)) ,
-                        min_logodds(OccupancyPatch::logodds(min_probability)),
-                        uncertainty_threshold(uncertainty_threshold) {}
+namespace maps { namespace grid 
+{   
 
-    float hit_logodds;
-    float miss_logodds;
-    float occupied_logodds;
-    float free_space_logodds;
-    float max_logodds;
-    float min_logodds;
-    float uncertainty_threshold;
+class OccupancyPatch
+{
+    float log_odds;
+
+public:
+    OccupancyPatch(double initial_probability) : log_odds(logodds(initial_probability)) {}
+    OccupancyPatch(float initial_log_odds = 0.f) : log_odds(initial_log_odds) {}
+    virtual ~OccupancyPatch() {}
+
+    double getPropability() const
+    {
+	return probability(log_odds);
+    }
+
+    float getLogOdds() const
+    {
+	return log_odds;
+    }
+
+    bool isOccupied(double occupied_tresshold = 0.8) const
+    {
+	return probability(log_odds) >= occupied_tresshold;
+    }
+
+    bool isFreeSpace(double not_occupied_tresshold = 0.3) const
+    {
+	return probability(log_odds) < not_occupied_tresshold;
+    }
+
+    void updatePropability(double update_prob, double min_prob = 0.1192, double max_prob = 0.971)
+    {
+	updateLogOdds(logodds(update_prob), logodds(min_prob), logodds(max_prob));
+    }
+
+    void updateLogOdds(float update_logodds, float min = -2.f, float max = 3.5f)
+    {
+	log_odds += update_logodds;
+	if(log_odds < min)
+	    log_odds = min;
+	else if(log_odds > max)
+	    log_odds = max;
+    }
+
+    bool operator==(const OccupancyPatch& other) const
+    {
+	return this == &other;
+    }
+
+    // compute log-odds from probability
+    static inline float logodds(double probability)
+    {
+	return (float)log(probability / (1.0 - probability));
+    }
+
+    // compute probability from log-odds
+    static inline double probability(double logodds)
+    {
+	return 1.0 - ( 1.0 / (1.0 + exp(logodds)));
+    }
 
 protected:
+
     /** Grants access to boost serialization */
     friend class boost::serialization::access;
 
@@ -64,14 +106,11 @@ protected:
     template <typename Archive>
     void serialize(Archive &ar, const unsigned int version)
     {
-        ar & BOOST_SERIALIZATION_NVP(hit_logodds);
-        ar & BOOST_SERIALIZATION_NVP(miss_logodds);
-        ar & BOOST_SERIALIZATION_NVP(occupied_logodds);
-        ar & BOOST_SERIALIZATION_NVP(free_space_logodds);
-        ar & BOOST_SERIALIZATION_NVP(max_logodds);
-        ar & BOOST_SERIALIZATION_NVP(min_logodds);
-        ar & BOOST_SERIALIZATION_NVP(uncertainty_threshold);
+	ar & BOOST_SERIALIZATION_NVP(log_odds);
     }
 };
+	
+}  //namespace grid
+}  //namespace maps
 
-}}
+#endif  //__MAPS_OCCUPANCY_PATCH_HPP_
