+++ src/grid/SurfacePatches.hpp
@@ -30,7 +30,6 @@
 #include <numeric/PlaneFitting.hpp>
 
 #include <base/Eigen.hpp>
-#include <base/Float.hpp>
 #include <base-logging/Logging.hpp>
 
 #include <Eigen/Core>
@@ -550,148 +549,6 @@ protected:
     BOOST_SERIALIZATION_SPLIT_MEMBER()
 }; // SurfacePatch<MLSConfig::PRECALCULATED>
 
-
-class OccupancyPatch
-{
-    float log_odds;
-
-public:
-    OccupancyPatch(double initial_probability) : log_odds(logodds(initial_probability)) {}
-    OccupancyPatch(float initial_log_odds = 0.f) : log_odds(initial_log_odds) {}
-    virtual ~OccupancyPatch() {}
-
-    double getPropability() const
-    {
-        return probability(log_odds);
-    }
-
-    float getLogOdds() const
-    {
-        return log_odds;
-    }
-
-    bool isOccupied(double occupied_tresshold = 0.8) const
-    {
-        return probability(log_odds) >= occupied_tresshold;
-    }
-
-    bool isFreeSpace(double not_occupied_tresshold = 0.3) const
-    {
-        return probability(log_odds) < not_occupied_tresshold;
-    }
-
-    void updatePropability(double update_prob, double min_prob = 0.1192, double max_prob = 0.971)
-    {
-        updateLogOdds(logodds(update_prob), logodds(min_prob), logodds(max_prob));
-    }
-
-    void updateLogOdds(float update_logodds, float min = -2.f, float max = 3.5f)
-    {
-        log_odds += update_logodds;
-        if(log_odds < min)
-            log_odds = min;
-        else if(log_odds > max)
-            log_odds = max;
-    }
-
-    bool operator==(const OccupancyPatch& other) const
-    {
-        return this == &other;
-    }
-
-    // compute log-odds from probability
-    static inline float logodds(double probability)
-    {
-        return (float)log(probability / (1.0 - probability));
-    }
-
-    // compute probability from log-odds
-    static inline double probability(double logodds)
-    {
-        return 1.0 - ( 1.0 / (1.0 + exp(logodds)));
-    }
-
-protected:
-
-    /** Grants access to boost serialization */
-    friend class boost::serialization::access;
-
-    /** Serializes the members of this class*/
-    template <typename Archive>
-    void serialize(Archive &ar, const unsigned int version)
-    {
-        ar & BOOST_SERIALIZATION_NVP(log_odds);
-    }
-};
-
-class TSDFPatch
-{
-    float distance;
-    float var;
-
-    template <class T> static inline void kalman_update( T& mean, T& var, T m_mean, T m_var )
-    {
-        float gain = var / (var + m_var);
-        if( gain != gain )
-            gain = 0.5f; // this happens when both vars are 0.
-        mean = mean + gain * (m_mean - mean);
-        var = (1.0f - gain) * var;
-    }
-
-public:
-    TSDFPatch() : distance(base::NaN<float>()), var(1.f) {}
-    TSDFPatch(float distance, float var) : distance(distance), var(var) {}
-    virtual ~TSDFPatch() {}
-
-    void update(float distance, float var, float truncation = 1.f, float min_var = 0.001f)
-    {
-        if(base::isNaN<float>(this->distance))
-            this->distance = distance;
-
-        kalman_update(this->distance, this->var, distance, var);
-
-        if(this->var < min_var)
-            this->var = min_var;
-
-        if(distance > truncation)
-            distance = truncation;
-        else if(distance < -truncation)
-            distance = -truncation;
-    }
-
-    float getDistance() const
-    {
-        return distance;
-    }
-
-    float getVariance() const
-    {
-        return var;
-    }
-
-    float getStandardDeviation() const
-    {
-        return std::sqrt(var);
-    }
-
-    bool operator==(const TSDFPatch& other) const
-    {
-        return this == &other;
-    }
-
-protected:
-
-    /** Grants access to boost serialization */
-    friend class boost::serialization::access;
-
-    /** Serializes the members of this class*/
-    template <typename Archive>
-    void serialize(Archive &ar, const unsigned int version)
-    {
-        ar & BOOST_SERIALIZATION_NVP(distance);
-        ar & BOOST_SERIALIZATION_NVP(var);
-    }
-};
 
 
 /**
